Update importer to handle large vertex lists
The supcom meshes store faces in a short int, and that can overflow.
When overflowing the file itself doesnt care, but blender does. As a
hack we conver the short int overflows by adding the limit to them
twice, which results in the correct vertex ID. its a bit of a mess but
it works.