Update vertex handling on import 2.80
The supcom meshes store faces in a short int, and that can overflow.
When overflowing the file itself doesnt care, but blender does. As a
hack we convert the short int overflows by adding the limit to them
twice, which results in the correct vertex ID. its a bit of a mess but
it works.