Update Exporter 2.79 to version 0.5.4
Now the exporter has support for many more bones, 256 used (with
geometry) and 256^2 unused. As best i can tell, this is the maximum the
scm file format allows.

Note: all the parents of used bones also count as used, even if they are
not used themselves, and therefore also fall in the 256 limit.