Update 2.80 Exporter to version 5.6
This is a pretty big deal.

Quad faces are now supported, and there is (once again) proper error
detection for Ngons, which arent supported (yet, growth mindset)

Additionally, sharp edges no longer need to be split to show up as sharp
in the game.

This should make it a lot easier to use the plugin now.