Update Exporter to 5.8
This adds support for ngons! Also it should be a tiny bit faster.

That said though it is still recommended to use edge splitting and
triangulation since there are a couple of minor bugs related to model
efficiency, so your model will be a bit more compact if you do that.